Submitted code for Riddler Express 11/16/18
Monte Carlo simulation of chess championship problem of how likely a player is to win a match with a variable number of games per match.

Problem statement:
The World Chess Championship is underway. It is a 12-game match between the world’s top two grandmasters. Many chess fans feel that 12 games is far too short for a biennial world championship match, allowing too much variance.

Say one of the players is better than his opponent to the degree that he wins 20 percent of all games, loses 15 percent of games and that 65 percent of games are drawn. Wins at this match are worth 1 point, draws a half-point for each player, and losses 0 points. In a 12-game match, the first player to 6.5 points wins.

What are the chances the better player wins a 12-game match? How many games would a match have to be in order to give the better player a 75 chance of winning the match outright? A 90 percent chance? A 99 percent chance?

Answers:
2 games: ~52% chance of better plaer winning.
Need to play at least 14 games per match to have >= 75% chance of better player winning (~84% chance in 14 games per match).
Need to play at least 18 games per match to have >=99% chance of better player winning (~99.5% chance in 18 games per match).